refact: organize create unit system
- uses  the settler API to create a settler
