+++ Assets/Scripts/Game/Unit/Systems/CreateUnitSystem.cs
@@ -1,0 +1,40 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Unity.Entities;
+using Unity.Jobs;
+using Unity.Transforms;
+using Unity.Mathematics;
+
+using Hex;
+
+namespace Game
+{
+    public class CreateUnitSystem : SystemBase
+    {
+        EndSimulationEntityCommandBufferSystem barrier => World.GetOrCreateSystem<EndSimulationEntityCommandBufferSystem>();
+
+        protected override void OnUpdate()
+        {
+            EntityCommandBuffer ecb = barrier.CreateCommandBuffer();
+            SettlerPrefab settlerPrefab = GetSingleton<SettlerPrefab>();
+
+            Entities
+                .WithoutBurst()
+                .ForEach((
+                    Entity entity,
+                    int entityInQueryIndex,
+                    in CommandCreateUnitComponent createUnitCommand
+                ) => {
+                    int3 coordinates = createUnitCommand.Coordinates;
+
+                    SettlerEntity.Create(ecb, settlerPrefab, coordinates);
+
+                    ecb.DestroyEntity(entity);
+                })
+                .Run();
+
+            barrier.AddJobHandleForProducer(this.Dependency);
+        }
+    }
+}
